+++ packages/wrangler/src/deployment-bundle/bundle.ts
@@ -1,5 +1,4 @@
 import * as fs from "node:fs";
-import { builtinModules } from "node:module";
 import * as path from "node:path";
 import NodeGlobalsPolyfills from "@esbuild-plugins/node-globals-polyfill";
 import NodeModulesPolyfills from "@esbuild-plugins/node-modules-polyfill";
@@ -7,7 +6,12 @@ import * as esbuild from "esbuild";
 import tmp from "tmp-promise";
 import createModuleCollector from "../module-collection";
 import { getBasePath } from "../paths";
-import { dedent } from "../utils/dedent";
+import { applyMiddlewareLoaderFacade } from "./apply-middleware";
+import {
+	isBuildFailure,
+	rewriteNodeCompatBuildFailure,
+} from "./build-failures";
+import { dedupeModulesByName } from "./dedupe-modules";
 import { getEntryPointFromMetafile } from "./entry-point-from-metafile";
 import { cloudflareInternalPlugin } from "./esbuild-plugins/cloudflare-internal";
 import { configProviderPlugin } from "./esbuild-plugins/config-provider";
@@ -16,6 +20,7 @@ import type { Config } from "../config";
 import type { DurableObjectBindings } from "../config/environment";
 import type { WorkerRegistry } from "../dev-registry";
 import type { ModuleCollector } from "../module-collection";
+import type { MiddlewareLoader } from "./apply-middleware";
 import type { Entry } from "./entry";
 import type { CfModule } from "./worker";
 
@@ -45,78 +50,39 @@ export type BundleResult = {
 	moduleCollector: ModuleCollector | undefined;
 };
 
-export type StaticAssetsConfig =
-	| (Config["assets"] & {
-			bypassCache: boolean | undefined;
-	  })
-	| undefined;
-
-/**
- * When applying the middleware facade for service workers, we need to inject
- * some code at the top of the final output bundle. Applying an inject too early
- * will allow esbuild to reorder the code. Additionally, we need to make sure
- * user code is bundled in the final esbuild step with `watch` correctly
- * configured, so code changes are detected.
- *
- * This type is used as the return type for the `MiddlewareFn` type representing
- * a facade-applying function. Returned injects should be injected with the
- * final esbuild step.
- */
-type EntryWithInject = Entry & { inject?: string[] };
-
-/**
- * RegExp matching against esbuild's error text when it is unable to resolve
- * a Node built-in module. If we detect this when node_compat is disabled,
- * we'll rewrite the error to suggest enabling it.
- */
-const nodeBuiltinResolveErrorText = new RegExp(
-	'^Could not resolve "(' +
-		builtinModules.join("|") +
-		"|" +
-		builtinModules.map((module) => "node:" + module).join("|") +
-		')"$'
-);
-
-/**
- * Returns true if the passed value looks like an esbuild BuildFailure object
- */
-export function isBuildFailure(err: unknown): err is esbuild.BuildFailure {
-	return (
-		typeof err === "object" &&
-		err !== null &&
-		"errors" in err &&
-		"warnings" in err
-	);
-}
-
-/**
- * Rewrites esbuild BuildFailures for failing to resolve Node built-in modules
- * to suggest enabling Node compat as opposed to `platform: "node"`.
- */
-export function rewriteNodeCompatBuildFailure(
-	errors: esbuild.Message[],
-	forPages = false
-) {
-	for (const error of errors) {
-		const match = nodeBuiltinResolveErrorText.exec(error.text);
-		if (match !== null) {
-			const issue = `The package "${match[1]}" wasn't found on the file system but is built into node.`;
-
-			const instructionForUser = `${
-				forPages
-					? 'Add the "nodejs_compat" compatibility flag to your Pages project'
-					: 'Add "node_compat = true" to your wrangler.toml file'
-			} to enable Node.js compatibility.`;
-
-			error.notes = [
-				{
-					location: null,
-					text: `${issue}\n${instructionForUser}`,
-				},
-			];
-		}
-	}
-}
+export type BundleOptions = {
+	// When `bundle` is set to false, we apply shims to the Worker, but won't pull in any imports
+	bundle?: boolean;
+	serveAssetsFromWorker: boolean;
+	assets?: Config["assets"];
+	bypassAssetCache?: boolean;
+	doBindings: DurableObjectBindings;
+	jsxFactory?: string;
+	jsxFragment?: string;
+	entryName?: string;
+	rules: Config["rules"];
+	watch?: boolean;
+	tsconfig?: string;
+	minify?: boolean;
+	legacyNodeCompat?: boolean;
+	nodejsCompat?: boolean;
+	define: Config["define"];
+	checkFetch: boolean;
+	services?: Config["services"];
+	workerDefinitions?: WorkerRegistry;
+	targetConsumer: "dev" | "deploy";
+	testScheduled?: boolean;
+	inject?: string[];
+	loader?: Record<string, string>;
+	sourcemap?: esbuild.CommonOptions["sourcemap"];
+	plugins?: esbuild.Plugin[];
+	additionalModules?: CfModule[];
+	// TODO: Rip these out https://github.com/cloudflare/workers-sdk/issues/2153
+	disableModuleCollection?: boolean;
+	isOutfile?: boolean;
+	forPages?: boolean;
+	local: boolean;
+};
 
 /**
  * Generate a bundle for the worker identified by the arguments passed in.
@@ -124,40 +90,7 @@ export function rewriteNodeCompatBuildFailure(
 export async function bundleWorker(
 	entry: Entry,
 	destination: string,
-	options: {
-		// When `bundle` is set to false, we apply shims to the Worker, but won't pull in any imports
-		bundle?: boolean;
-		serveAssetsFromWorker: boolean;
-		assets?: StaticAssetsConfig;
-		doBindings: DurableObjectBindings;
-		jsxFactory?: string;
-		jsxFragment?: string;
-		entryName?: string;
-		rules: Config["rules"];
-		watch?: boolean;
-		tsconfig?: string;
-		minify?: boolean;
-		legacyNodeCompat?: boolean;
-		nodejsCompat?: boolean;
-		define: Config["define"];
-		checkFetch: boolean;
-		services?: Config["services"];
-		workerDefinitions?: WorkerRegistry;
-		targetConsumer: "dev" | "deploy";
-		testScheduled?: boolean;
-		inject?: string[];
-		loader?: Record<string, string>;
-		sourcemap?: esbuild.CommonOptions["sourcemap"];
-		plugins?: esbuild.Plugin[];
-		additionalModules?: CfModule[];
-		// TODO: Rip these out https://github.com/cloudflare/workers-sdk/issues/2153
-		disableModuleCollection?: boolean;
-		isOutfile?: boolean;
-		forPages?: boolean;
-		local: boolean;
-	}
-): Promise<BundleResult> {
-	const {
+	{
 		bundle = true,
 		serveAssetsFromWorker,
 		doBindings,
@@ -170,8 +103,10 @@ export async function bundleWorker(
 		minify,
 		legacyNodeCompat,
 		nodejsCompat,
+		define,
 		checkFetch,
 		assets,
+		bypassAssetCache,
 		workerDefinitions,
 		services,
 		targetConsumer,
@@ -185,8 +120,8 @@ export async function bundleWorker(
 		forPages,
 		additionalModules = [],
 		local,
-	} = options;
-
+	}: BundleOptions
+): Promise<BundleResult> {
 	// We create a temporary directory for any one-off files we
 	// need to create. This is separate from the main build
 	// directory (`destination`).
@@ -195,7 +130,9 @@ export async function bundleWorker(
 	// without symlinks, otherwise `esbuild` will generate invalid source maps.
 	const tmpDirPath = fs.realpathSync(unsafeTmpDir.path);
 
-	const entryDirectory = path.dirname(entry.file);
+	const entryFile = entry.file;
+
+	const entryDirectory = path.dirname(entryFile);
 	let moduleCollector = createModuleCollector({
 		wrangler1xlegacyModuleReferences: {
 			rootDirectory: entryDirectory,
@@ -204,7 +141,7 @@ export async function bundleWorker(
 					.readdirSync(entryDirectory, { withFileTypes: true })
 					.filter(
 						(dirEntry) =>
-							dirEntry.isFile() && dirEntry.name !== path.basename(entry.file)
+							dirEntry.isFile() && dirEntry.name !== path.basename(entryFile)
 					)
 					.map((dirEnt) => dirEnt.name)
 			),
@@ -222,26 +159,7 @@ export async function bundleWorker(
 		};
 	}
 
-	// In dev, we want to patch `fetch()` with a special version that looks
-	// for bad usages and can warn the user about them; so we inject
-	// `checked-fetch.js` to do so. However, with yarn 3 style pnp,
-	// we need to extract that file to an accessible place before injecting
-	// it in, hence this code here.
-
-	const checkedFetchFileToInject = path.join(tmpDirPath, "checked-fetch.js");
-
-	if (checkFetch && !fs.existsSync(checkedFetchFileToInject)) {
-		fs.mkdirSync(tmpDirPath, {
-			recursive: true,
-		});
-		fs.writeFileSync(
-			checkedFetchFileToInject,
-			fs.readFileSync(path.resolve(getBasePath(), "templates/checked-fetch.js"))
-		);
-	}
-
-	// At this point, we take the opportunity to "wrap" any input workers
-	// with any extra functionality we may want to add.
+	// At this point, we take the opportunity to "wrap" the worker with middleware.
 	const middlewareToLoad: MiddlewareLoader[] = [];
 
 	if (targetConsumer === "dev" && !!testScheduled) {
@@ -272,6 +190,7 @@ export async function bundleWorker(
 			path: "templates/middleware/middleware-miniflare3-json-error.ts",
 		});
 	}
+
 	if (serveAssetsFromWorker) {
 		middlewareToLoad.push({
 			name: "serve-static-assets",
@@ -284,7 +203,7 @@ export async function bundleWorker(
 						? {
 								browserTTL:
 									assets.browser_TTL || 172800 /* 2 days: 2* 60 * 60 * 24 */,
-								bypassCache: assets.bypassCache,
+								bypassCache: bypassAssetCache,
 						  }
 						: {},
 			},
@@ -314,8 +233,8 @@ export async function bundleWorker(
 		});
 	}
 
-	// If using watch, build result will not be returned
-	// This plugin will retreive the build result on the first build
+	// If using watch, build result will not be returned.
+	// This plugin will retrieve the build result on the first build.
 	let initialBuildResult: (result: esbuild.BuildResult) => void;
 	const initialBuildResultPromise = new Promise<esbuild.BuildResult>(
 		(resolve) => (initialBuildResult = resolve)
@@ -328,27 +247,64 @@ export async function bundleWorker(
 	};
 
 	const inject: string[] = injectOption ?? [];
-	if (checkFetch) inject.push(checkedFetchFileToInject);
-	let inputEntry: EntryWithInject = entry;
+
+	if (checkFetch) {
+		// In dev, we want to patch `fetch()` with a special version that looks
+		// for bad usages and can warn the user about them; so we inject
+		// `checked-fetch.js` to do so. However, with yarn 3 style pnp,
+		// we need to extract that file to an accessible place before injecting
+		// it in, hence this code here.
+
+		const checkedFetchFileToInject = path.join(tmpDirPath, "checked-fetch.js");
+
+		if (checkFetch && !fs.existsSync(checkedFetchFileToInject)) {
+			fs.mkdirSync(tmpDirPath, {
+				recursive: true,
+			});
+			fs.writeFileSync(
+				checkedFetchFileToInject,
+				fs.readFileSync(
+					path.resolve(getBasePath(), "templates/checked-fetch.js")
+				)
+			);
+		}
+
+		inject.push(checkedFetchFileToInject);
+	}
+
 	if (
 		middlewareToLoad.length > 0 ||
 		process.env.EXPERIMENTAL_MIDDLEWARE === "true"
 	) {
-		inputEntry = await applyMiddlewareLoaderFacade(
+		const result = await applyMiddlewareLoaderFacade(
 			entry,
 			tmpDirPath,
 			middlewareToLoad,
 			doBindings
 		);
-		if (inputEntry.inject !== undefined) inject.push(...inputEntry.inject);
+		entry = result.entry;
+
+		/**
+		 * When applying the middleware facade for service workers, we need to inject
+		 * some code at the top of the final output bundle. Applying an inject too early
+		 * will allow esbuild to reorder the code. Additionally, we need to make sure
+		 * user code is bundled in the final esbuild step with `watch` correctly
+		 * configured, so code changes are detected.
+		 *
+		 * This type is used as the return type for the `MiddlewareFn` type representing
+		 * a facade-applying function. Returned injects should be injected with the
+		 * final esbuild step.
+		 */
+		inject.push(...(result.inject ?? []));
 	}
 
 	const buildOptions: esbuild.BuildOptions & { metafile: true } = {
-		entryPoints: [inputEntry.file],
+		// Don't use entryFile here as the file may have been changed when applying the middleware
+		entryPoints: [entry.file],
 		bundle,
 		absWorkingDir: entry.directory,
 		outdir: destination,
-		entryNames: entryName || path.parse(entry.file).name,
+		entryNames: entryName || path.parse(entryFile).name,
 		...(isOutfile
 			? {
 					outdir: undefined,
@@ -373,7 +329,7 @@ export async function bundleWorker(
 				// when we do a build of wrangler. (re: https://github.com/cloudflare/workers-sdk/issues/1477)
 				"process.env.NODE_ENV": `"${process.env["NODE_ENV" + ""]}"`,
 				...(legacyNodeCompat ? { global: "globalThis" } : {}),
-				...options.define,
+				...define,
 			},
 		}),
 		loader: {
@@ -429,7 +385,7 @@ export async function bundleWorker(
 		throw e;
 	}
 
-	const entryPoint = getEntryPointFromMetafile(entry.file, result.metafile);
+	const entryPoint = getEntryPointFromMetafile(entryFile, result.metafile);
 	const bundleType = entryPoint.exports.length > 0 ? "esm" : "commonjs";
 
 	const sourceMapPath = Object.keys(result.metafile.outputs).filter((_path) =>
@@ -471,167 +427,4 @@ export async function bundleWorker(
 		},
 		moduleCollector,
 	};
-}
-
-/**
- * A facade that acts as a "middleware loader".
- * Instead of needing to apply a facade for each individual middleware, this allows
- * middleware to be written in a more traditional manner and then be applied all
- * at once, requiring just two esbuild steps, rather than 1 per middleware.
- */
-interface MiddlewareLoader {
-	name: string;
-	path: string;
-	// Each config item will be provided as a virtual module at config:middleware/${name}
-	// where `name` is a key property on the `config` record.
-	config?: Record<string, unknown>;
-}
-
-async function applyMiddlewareLoaderFacade(
-	entry: Entry,
-	tmpDirPath: string,
-	middleware: MiddlewareLoader[], // a list of paths to middleware files
-	doBindings: DurableObjectBindings
-): Promise<EntryWithInject> {
-	// Firstly we need to insert the middleware array into the project,
-	// and then we load the middleware - this insertion and loading is
-	// different for each format.
-
-	// We need to import each of the middlewares, so we need to generate a
-	// random, unique identifier that we can use for the import.
-	// Middlewares are required to be default exports so we can import to any name.
-	const middlewareIdentifiers = middleware.map((m, index) => [
-		`__MIDDLEWARE_${index}__`,
-		path.resolve(getBasePath(), m.path),
-	]);
-
-	const dynamicFacadePath = path.join(
-		tmpDirPath,
-		"middleware-insertion-facade.js"
-	);
-	const imports = middlewareIdentifiers
-		.map(
-			([id, middlewarePath]) =>
-				/*javascript*/ `import * as ${id} from "${prepareFilePath(
-					middlewarePath
-				)}";`
-		)
-		.join("\n");
-
-	const middlewareFns = middlewareIdentifiers.map(([m]) => `${m}.default`);
-
-	if (entry.format === "modules") {
-		const middlewareWrappers = middlewareIdentifiers
-			.map(([m]) => `${m}.wrap`)
-			.join(",");
-
-		const durableObjects = doBindings
-			// Don't shim anything not local to this worker
-			.filter((b) => !b.script_name)
-			// Reexport the DO classnames
-			.map(
-				(b) =>
-					/*javascript*/ `export const ${b.class_name} = maskDurableObjectDefinition(OTHER_EXPORTS.${b.class_name});`
-			)
-			.join("\n");
-		await fs.promises.writeFile(
-			dynamicFacadePath,
-			dedent/*javascript*/ `
-				import worker, * as OTHER_EXPORTS from "${prepareFilePath(entry.file)}";
-				${imports}
-				const envWrappers = [${middlewareWrappers}].filter(Boolean);
-				const facade = {
-					...worker,
-					envWrappers,
-					middleware: [
-						${middlewareFns.join(",")},
-            ...(worker.middleware ? worker.middleware : []),
-					].filter(Boolean)
-				}
-				export * from "${prepareFilePath(entry.file)}";
-
-				const maskDurableObjectDefinition = (cls) =>
-					class extends cls {
-						constructor(state, env) {
-							let wrappedEnv = env
-							for (const wrapFn of envWrappers) {
-								wrappedEnv = wrapFn(wrappedEnv)
-							}
-							super(state, wrappedEnv);
-						}
-					};
-				${durableObjects}
-
-				export default facade;
-			`
-		);
-
-		const targetPathLoader = path.join(
-			tmpDirPath,
-			"middleware-loader.entry.ts"
-		);
-		const loaderPath = path.resolve(
-			getBasePath(),
-			"templates/middleware/loader-modules.ts"
-		);
-
-		const baseLoader = await fs.promises.readFile(loaderPath, "utf-8");
-		const transformedLoader = baseLoader
-			.replaceAll("__ENTRY_POINT__", prepareFilePath(dynamicFacadePath))
-			.replace(
-				"./common",
-				prepareFilePath(
-					path.resolve(getBasePath(), "templates/middleware/common.ts")
-				)
-			);
-
-		await fs.promises.writeFile(targetPathLoader, transformedLoader);
-
-		return {
-			...entry,
-			file: targetPathLoader,
-		};
-	} else {
-		const loaderSwPath = path.resolve(
-			getBasePath(),
-			"templates/middleware/loader-sw.ts"
-		);
-
-		await fs.promises.writeFile(
-			dynamicFacadePath,
-			dedent/*javascript*/ `
-				import { __facade_registerInternal__ } from "${prepareFilePath(loaderSwPath)}";
-				${imports}
-				__facade_registerInternal__([${middlewareFns}])
-			`
-		);
-
-		return {
-			...entry,
-			inject: [dynamicFacadePath],
-		};
-	}
-}
-
-/**
- * Prefer modules towards the end of the array in the case of a collision by name.
- */
-export function dedupeModulesByName(modules: CfModule[]): CfModule[] {
-	return Object.values(
-		modules.reduce((moduleMap, module) => {
-			moduleMap[module.name] = module;
-			return moduleMap;
-		}, {} as Record<string, CfModule>)
-	);
-}
-/**
- * Process the given file path to ensure it will work on all OSes.
- *
- * Windows paths contain backslashes, which are taken to be escape characters
- * when inserted directly into source code.
- * This function will escape these backslashes to make sure they work in all OSes.
- *
- */
-function prepareFilePath(filePath: string): string {
-	return JSON.stringify(filePath).slice(1, -1);
 }
